# Micro-optimization: Make ArgKind a regular class instead of enum