@ilevkivskyi 2025-07-17
@sterliakov I think we can optimize enums in general in mypyc (e.g. use some static pointers for enum values, use pointer identity comparison for `==`, and inline enum references when compiling, i.e. `SomeEnum.some_val` will become `_CPy_SomeEnum_some_val`). Most likely the perf benefit is from not having to go through attribute access logic, but TBH I didn't really think much about this.

That said I agree having optimized enums in mypyc is IMO worth it.