@JukkaL 2025-07-18
@sterliakov @ilevkivskyi I created a mypyc issue about this: https://github.com/mypyc/mypyc/issues/1121

Calling methods of enum classes is the biggest bottleneck that I can see. Accessing enum values is already optimized.